Cleanup, correction and speed optimization due to the last restructurization. Most int -> short array copy loops are now eliminated.
The color names now got the final prefix G_ (GEM color) instead
the not 100% correct OC_ (object color) prefix.

+++ gemlib/n_vqt_xfntinfo.c
@@ -3,9 +3,9 @@
 
 
 short
-vqt_xfntinfo (short handle, short flags, short id, short index, XFNT_INFO * info)
+vqt_xfntinfo (short handle, short flags, short id, short index, XFNT_INFO *info)
 {
-	info->size = (long) sizeof (XFNT_INFO);
+	info->size = sizeof (XFNT_INFO);
 	vdi_intin[0] = flags;
 	vdi_intin[1] = id;
 	vdi_intin[2] = index;
